Remove path method from PathBuilder
