+++ src/patterns/paths/utils.ts
@@ -9,24 +9,6 @@ import { Path } from "./Path";
 
 
 /**
- * Create a {@link Path} from the property provided.
- *
- * @param container The general container with the {@link IRIResolver}
- * that will be used to resolve a string property.
- * @param property The property to be converted into a Path.
- */
-export function parseProperty( container:Container<any>, property:Resource | "a" | string ):Path<IRIToken | "a"> {
-	const targetToken:IRIToken | "a" = getPropertyToken( container, property );
-
-	const newContainer:Container<IRIToken | "a"> = new Container( {
-		iriResolver: container.iriResolver,
-		targetToken,
-	} );
-	return Path.createFrom( newContainer, {} );
-}
-
-
-/**
  * Obtain the token representation from the property provided.
  *
  * @param container The general container with the {@link IRIResolver}
